# Conflicts:
#	pkg/dto/storage_s3.go

+++ internal/server/handlers/config_cluster_test.go
@@ -25,7 +25,7 @@ func testSeedNode() dto.SeedNode {
 
 func testConfigCluster() *dto.AerospikeCluster {
 	label := "label"
-	timeout := int32(10)
+	timeout := 10
 	useAlternate := false
 	queueSize := 1
 	return &dto.AerospikeCluster{
